util| move the github bits out into their own file
